LE-26 Obtain a JSESSIONID from the back-end as part of Logging in successfully
This is more than the JSESSIONID, but it allows the REST API to
work and that was the primary goal of this ticket.  This has switched
over to token-based authentication. Changes:
* Adds OAuth endpoint for use with the Google API for Authentication.
* Adds CORS headers to the CluerideSessionFilter which checks
sessions and establishes them if sufficient credentials are provided.
* Allows the establishment of a session to generate a token that will
be used to authenticate on each request.

Remaining work:
* Create a genuine token to be used.
* Actually pay attention to that token.
* Further modularization of the filters (if called for).
* There are also a number of improvements that can be made incorporating
JAX-RS and Jersey into this solution.