CA-367 Prepare for, implement and test Write access to a wp_users table via Hibernate
- Adds Store and Service layers for a new WpUser class representing WordPress Users.
This class is (mostly) independent of BadgeOS.

Full creation of the user account requires population of "meta" records which is not performed by
this store, but may be performed by the Service layer talking with a separate Store.

Learned that the persistence language expects entity names and if we have two different entities
going against the same table, we need to expressly map to that table using two different entity
names.